# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App - Changed
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths 
basePath: "/"
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  #/hello:
  /:  
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /soma:
    x-swagger-router-controller: soma
    post:
      summary: Executes a Fraud Check
      description: >
        text
      parameters:
        - in: body
          name: body
          description: description test.
          required: true
          schema:
            $ref: "#/definitions/SomaRequest"
      #tags:
      #  - Client
      #  - Fraud Check
      responses:
        "201":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SomaResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users:
    x-swagger-router-controller: users route controller
    get:
      description: Return the list of users
      operationId: Return Users
      responses:
        "200":
          description: Return the list of users
          schema:
            $ref: "#/definitions/usersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{userId}:
    x-swagger-router-controller: users route controller
    get:
      description: Return the content of a specific user
      parameters:
        - name: userId
          in: path
          description: User Id.
          required: true
          type: number
      operationId: Return User content
      responses:
        "200":
          description: Return the list of users
          schema:
            $ref: "#/definitions/usersResponseArrayItems"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update a specific user
      parameters:
        - name: userId
          in: path
          description: User Id.
          required: true
          type: number
        - in: body
          name: body
          description: Data to create an user.
          required: true
          schema:
            $ref: "#/definitions/userRequest"
      operationId: Update User content
      responses:
        "200":
          description: User updated
          schema:
            $ref: "#/definitions/messageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Delete a specific user
      parameters:
        - name: userId
          in: path
          description: User Id.
          required: true
          type: number
      operationId: Return User content
      responses:
        "200":
          description: User deleted
          schema:
            $ref: "#/definitions/messageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/:
    x-swagger-router-controller: users route controller
    post:
      description: create an user
      operationId: Create an User account
      parameters:
        - in: body
          name: body
          description: Data to create an user.
          required: true
          schema:
            $ref: "#/definitions/userRequest"
      responses:
        "200":
          description: Return the data of a created user and an userId
          schema:
            $ref: "#/definitions/usersResponseArrayItems"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/deleteAll:
    x-swagger-router-controller: users route controller
    delete:
      description: This URL will delete all the users
      operationId: Delete All Users
      responses:
        "200":
          description: Will delete all the users of the system
          schema:
            $ref: "#/definitions/messageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SomaRequest:
    required:
      - a
      - b    
    type: object
    properties:
      a:
        type: number
        minimum: 1
      b:
        type: number
        minimum: 1
  SomaResponse:
    required:
      - resultado
      - valores
    type: object
    properties:
      resultado:
        type: number
        minimum: 1
      valores:
        type: array
        items:
          type: number
  usersResponse:
    type: array
    items:
      $ref: "#/definitions/usersResponseArrayItems"
  usersResponseArrayItems:
    type: object
    properties:
      user:
        type: object
        properties:
          username:
            type: string
          registry:
            type: string
      id:
        type: number
  usersResponseArrayUserItems:
    type: object
    properties:
      username:
        type: string
      registry:
        type: string
  userRequest:
    required:
      - username
      - registry    
    type: object
    properties:
      username:
        type: string
      registry:
        type: string
  messageResponse:
    required:
      - message
    properties:
      message:
        type: string